tests: merge append_parity_test.go to {append_encode,fuzz}_test.go
