+++ fuzz_test.go
@@ -5,8 +5,10 @@ package protocol
 
 import (
 	"bytes"
+	"math"
 	"testing"
 
+	"github.com/go-json-experiment/json"
 	"github.com/go-json-experiment/json/jsontext"
 )
 
@@ -122,4 +124,63 @@ func canonicalize(t *testing.T, b []byte) []byte {
 		t.Fatalf("canonicalize %s: %v", b, err)
 	}
 	return v
+}
+
+// FuzzAppendEncodeParity proves the generated byte-append writers emit
+// byte-identical output to the retained streaming/reflection oracle for every
+// value the byte decoders accept. Error outcomes must agree as well.
+func FuzzAppendEncodeParity(f *testing.F) {
+	for _, seed := range wellFormedInputs(f) {
+		f.Add(seed)
+	}
+
+	f.Fuzz(func(t *testing.T, data []byte) {
+		for _, newDst := range appendParityTargets() {
+			v := newDst()
+			if err := Unmarshal(data, v); err != nil {
+				continue
+			}
+			got, gerr := Marshal(v)
+			want, werr := json.Marshal(v, wireOptions)
+			if (gerr != nil) != (werr != nil) {
+				t.Fatalf("error parity for %T: append=%v oracle=%v", v, gerr, werr)
+			}
+			if gerr != nil {
+				continue
+			}
+			if !bytes.Equal(got, want) {
+				t.Fatalf("byte parity for %T:\n append=%s\n oracle=%s", v, got, want)
+			}
+		}
+	})
+}
+
+// FuzzAppendFloat64Parity proves appendFloat64JSON matches the jsontext
+// number formatting for every float64 bit pattern, including the non-finite
+// error contract.
+func FuzzAppendFloat64Parity(f *testing.F) {
+	for _, seed := range []uint64{
+		0, math.Float64bits(math.Copysign(0, -1)), math.Float64bits(1),
+		math.Float64bits(0.5), math.Float64bits(1e-6), math.Float64bits(9.999e-7),
+		math.Float64bits(1e20), math.Float64bits(1e21), math.Float64bits(math.MaxFloat64),
+		math.Float64bits(math.SmallestNonzeroFloat64), math.Float64bits(-123456.789e-12),
+		math.Float64bits(math.NaN()), math.Float64bits(math.Inf(1)),
+	} {
+		f.Add(seed)
+	}
+
+	f.Fuzz(func(t *testing.T, bits uint64) {
+		v := math.Float64frombits(bits)
+		got, gerr := appendFloat64JSON(nil, v)
+		want, werr := json.Marshal(v, wireOptions)
+		if (gerr != nil) != (werr != nil) {
+			t.Fatalf("error parity for %x (%v): append=%v oracle=%v", bits, v, gerr, werr)
+		}
+		if gerr != nil {
+			return
+		}
+		if !bytes.Equal(got, want) {
+			t.Fatalf("float parity for %x (%v): append=%s oracle=%s", bits, v, got, want)
+		}
+	})
 }
